FIX: Product Fetch Before Delete
(cherry picked from commit 3cced245ee6eeef0fcd4a1a9e7d0c08e27997514)

+++ splash/src/Objects/Product/CRUDTrait.php
@@ -187,6 +187,11 @@ trait CRUDTrait
         // Load Product Combination
         $combination = VariantsManager::getProductCombination((int) $objectId);
         //====================================================================//
+        // Fetch Object
+        if ($object->fetch((int) $objectId) <= 0) {
+            return $this->catchDolibarrErrors($object);
+        }
+        //====================================================================//
         // Delete Object
         if ($object->delete($user) <= 0) {
             return $this->catchDolibarrErrors($object);
